Added stop database setup
