+++ pyproject.toml
@@ -18,6 +18,7 @@ dependencies = [
     "polars>=1.29.0",
     "xmlschema>=4.0.1",
     "jsonschema-default>=1.8.1",
+    "pyarrow>=20.0.0"
 ]
 keywords = ["Python", "API", "Flask", "OpenBus", "BusTracker"]
 classifiers = [
